Ipmlement shared Base History Projector
